% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_neighborfinder.R
\name{apply_NeighborFinder}
\alias{apply_NeighborFinder}
\title{Apply NeighborFinder on raw data}
\usage{
apply_NeighborFinder(
  data_with_annotation,
  object_of_interest,
  col_module_id,
  annotation_level,
  prev_level = 0.3,
  filtering_top = 20,
  .seed = NULL,
  ...
)
}
\arguments{
\item{data_with_annotation}{Dataframe. The abundance table merged with the module names. Required format: modules are the rows and samples are the columns. The first column must be the modules name (e.g. species), the second is the module ID (e.g. msp), and each subsequent column is a sample}

\item{object_of_interest}{String. The name of the bacteria or species of interest or a key word in the functional module definition}

\item{col_module_id}{String. The name of the column with the module names in annotation_table}

\item{annotation_level}{String. The name of the column with the level to be studied. Examples: species, genus, level_1}

\item{prev_level}{Numeric. The prevalence to be studied. Required format is decimal: 0.20 for 20\% of prevalence}

\item{filtering_top}{Numeric. The filtering top percentage to be studied. Required format is: 10 for top 10\%}

\item{.seed}{Integer. Top level RNG seed to control the generation of RNG seed for the inner loop or NULL if reproducibility is not required.}

\item{...}{Additional arguments passed on to \code{\link[=apply_NF_simple]{apply_NF_simple()}}}
}
\value{
Dataframe. Returns results after using apply_NeighborFinder(): for each module ID from 'object_of_interest', the names of their neighbors and the corresponding coefficients calculated by cv.glmnet()
}
\description{
Apply NeighborFinder on raw data
}
\examples{
data(data)
res_CRC_JPN <- apply_NeighborFinder(
  data$CRC_JPN[, 1:100],
  object_of_interest = "Escherichia coli",
  col_module_id = "msp_id",
  annotation_level = "species",
  .seed = 123
)
}
