% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_precision.R
\name{compute_precision}
\alias{compute_precision}
\title{Compute precision rate}
\usage{
compute_precision(true, detected)
}
\arguments{
\item{true}{List. The one of true neighbors}

\item{detected}{List. The one of detected neighbors}
}
\value{
Numeric. Returns the precision rate in \%
}
\description{
Compute precision rate
}
\examples{
compute_precision(c("a"), c("a", "b", "c"))
compute_precision(c("a","b"), c("a", "c"))
}
