% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_recall.R
\name{compute_recall}
\alias{compute_recall}
\title{Compute recall rate}
\usage{
compute_recall(true, detected)
}
\arguments{
\item{true}{List. The one of true neighbors}

\item{detected}{List. The one of detected neighbors}
}
\value{
Numeric. Returns the recall rate
}
\description{
Compute recall rate
}
\examples{
compute_recall(c("a"), c("a", "b", "c"))
compute_recall(c("a", "b"), c("a", "c"))
}
