% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvglm_to_coeffs_by_bact.R
\name{cvglm_to_coeffs_by_bact}
\alias{cvglm_to_coeffs_by_bact}
\title{Apply cv.glmnet() for a list of msps and for each prevalence level}
\usage{
cvglm_to_coeffs_by_bact(list_dfs, test_msp = identify_msp(), seed = NULL, ...)
}
\arguments{
\item{list_dfs}{List of dataframe. A normalized dataframe}

\item{test_msp}{List of string. The msp names}

\item{seed}{Numeric. The seed number, ensuring reproducibility}

\item{covar}{String or formula. Formula or the name of the column of the covariate in the metadata table. Note that "study_accession" is equivalent to ~study_accession}

\item{meta_df}{Dataframe. The dataframe giving metadata information}

\item{sample_col}{String. The name of the column in metadata indicating the sample names, it should be consistent with the colnames of 'df'}
}
\value{
Dataframe. Returns the msp, its detected neighbor and the corresponding coefficient
}
\description{
Apply cv.glmnet() for a list of msps and for each prevalence level
}
\examples{
data(data)
data(metadata)
#Simple example
normed_JPN <-norm_data(data$CRC_JPN, col_msp_id="msp_id", prev_list=c(0.20, 0.25, 0.30), type="fpkm")
neighbors_JPN<-cvglm_to_coeffs_by_bact(list_dfs=normed_JPN, test_msp=c("msp_0030","msp_0345"), seed=20232024)
#Example with covariate
normed_CHN <-norm_data(data$CRC_CHN, col_msp_id="msp_id", prev_list=c(0.20, 0.25, 0.30), type="fpkm")
neighbors_CHN<-cvglm_to_coeffs_by_bact(list_dfs=normed_CHN, test_msp=c("msp_0030","msp_0345"), seed=20232024, covar= ~ study_accession, meta_df=metadata$CRC_CHN, sample_col="secondary_sample_accession")
}
