% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-for-crc-patients.R
\docType{data}
\name{data}
\alias{data}
\title{data}
\format{
An object of class \code{list} of length 4.
}
\source{
https://entrepot.recherche.data.gouv.fr/dataset.xhtml?persistentId=doi:10.57745/7IVO3E
}
\usage{
data
}
\description{
#' @format A list of dataframes corresponding to abundance tables merges with taxonomic information.
\describe{
\item{ CRC_JPN }{  dataframe with only Japanese patients diagnosed with colorectal cancer }
\item{ CRC_CHN }{  dataframe with only Chinese patients diagnosed with colorectal cancer }
\item{ CRC_EUR }{  dataframe with only European patients diagnosed with colorectal cancer }
\item{ CRC_JPN_CHN_EUR }{  dataframe with patients diagnosed with colorectal cancer from the 3 previous countries }
}
}
\keyword{datasets}
