% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_msp.R
\name{identify_msp}
\alias{identify_msp}
\title{List the msps corresponding to a given species of interest}
\usage{
identify_msp(bact_of_interest = "prausnitzii", taxo, col_msp_id)
}
\arguments{
\item{bact_of_interest}{String. The name of the bacteria or species of interest}

\item{taxo}{Dataframe. The dataframe gathering the taxonomic correspondence information}

\item{col_msp_id}{String. The name of the column with the msp names in taxo}
}
\value{
List of string. The msps names of the corresponding species of interest
}
\description{
List the msps corresponding to a given species of interest
}
\examples{
df_taxo<-data.frame(msp_name=c("msp_1","msp_2","msp_3","msp_4"),
                    genus=c("One","One","One","Two"),
                    species=c("One bacteria","One bacterium L","One bacterium G","Two bact"))
identify_msp("bacterium", taxo=df_taxo, col_msp_id = "msp_name")
identify_msp("One", taxo=df_taxo, col_msp_id = "msp_name")
}
