% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersections_table.R
\name{intersections_table}
\alias{intersections_table}
\title{Display the intersection table summarizing the results from 2 or more datasets}
\usage{
intersections_table(
  res_list,
  threshold,
  annotation_table,
  col_module_id,
  annotation_level,
  object_of_interest
)
}
\arguments{
\item{res_list}{List of dataframes. The results from apply_NeighborFinder() on several datasets}

\item{threshold}{Numeric. Integer corresponding to the minimum number of datasets in which you want neighbors to have been found}

\item{annotation_table}{Dataframe. The dataframe gathering the taxonomic or functional module correspondence information}

\item{col_module_id}{String. The name of the column with the module names in annotation_table}

\item{annotation_level}{String. The name of the column with the level to be studied. Examples: species, genus, level_1}

\item{object_of_interest}{String. The name of the bacteria or species of interest or a key word in the functional module definition}
}
\value{
Dataframe. Table gathering the intersection of NeighborFinder results from several datasets. The column 'datasets' indicates the datasets in which the same neighbor has been found, the column 'intersections' indicates the number of datasets in which the same neighbor has been found
}
\description{
Display the intersection table summarizing the results from 2 or more datasets
}
\examples{
data(taxo)
data(result_example)
intersections_table(res_list = list(result_example$res_CRC_JPN, result_example$res_CRC_CHN, result_example$res_CRC_EUR), threshold = 2, taxo, col_module_id = "msp_id", annotation_level = "species", "Escherichia coli")
}
