% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclr.R
\name{mclr}
\alias{mclr}
\title{Modified central log ratio (mclr) transformation extracted from the SPRING package}
\usage{
mclr(dat, base = exp(1), tol = 1e-16, eps = NULL, atleast = 1)
}
\arguments{
\item{dat}{raw count data or compositional data (n by p) does not matter.}

\item{base}{exp(1) for natural log}

\item{tol}{tolerance for checking zeros}

\item{eps}{epsilon in eq (2) of the paper "Yoon, Gaynanova, Müller (2019), Frontiers in Genetics". positive shifts to all non-zero compositions. Refer to the paper for more details. eps = absolute value of minimum of log ratio counts plus c.}

\item{atleast}{default value is 1. Constant c which ensures all nonzero values to be strictly positive. default is 1.}
}
\value{
\code{mclr} returns a data matrix of the same dimension with input data matrix.
}
\description{
Modified central log ratio (mclr) transformation extracted from the SPRING package
}
\details{
This function implements the mclr normalization introduced in Yoon G, Gaynanova I and Müller CL (2019) Microbial Networks in SPRING - Semi-parametric Rank-Based Correlation and Partial Correlation Estimation for Quantitative Microbiome Data. Front. Genet. 10:516. \doi{10.3389/fgene.2019.00516}
}
