% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msp_to_bact.R
\name{msp_to_bact}
\alias{msp_to_bact}
\title{Correspondence between msp name and bacteria name}
\usage{
msp_to_bact(msp, taxo, col_msp_id = "id_mgs")
}
\arguments{
\item{msp}{String. The name of the msp, can be a single one or a list}

\item{taxo}{Dataframe. The dataframe gathering the taxonomic correspondence information}

\item{col_msp_id}{String. The name of the column with the msp names in taxo}
}
\value{
Dictionary. The name of the bacteria, can be a single one or a list
}
\description{
Correspondence between msp name and bacteria name
}
\examples{
df_taxo<-data.frame(msp_name=c("msp_1","msp_2","msp_3","msp_4"),
                    genus=c("A","B","C","D"),
                    species=c("A a","B a","C c","D b"))
# Correspondence for one specific msp
msp_to_bact("msp_1", taxo=df_taxo, col_msp_id = "msp_name")
# or for several msps
msp_to_bact(c("msp_1","msp_3","msp_4"), taxo=df_taxo, col_msp_id = "msp_name")
# and if one msp is repeated
msp_to_bact(c("msp_1","msp_1","msp_2"), taxo=df_taxo, col_msp_id = "msp_name")
}
