% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result-for-a-use-case-example.R
\docType{data}
\name{result_example}
\alias{result_example}
\title{result_example}
\format{
A list of dataframes corresponding to apply_NeighborFinder() results.
\describe{
\item{ res_CRC_JPN }{  edge table of Escherichia coli neighbors with only Japanese patients diagnosed with colorectal cancer }
\item{ res_CRC_CHN }{  edge table of Escherichia coli neighbors with only Chinese patients diagnosed with colorectal cancer }
\item{ res_CRC_EUR }{  edge table of Escherichia coli neighbors with only European patients diagnosed with colorectal cancer }
}
}
\usage{
result_example
}
\description{
result_example
}
\keyword{datasets}
