% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo-data.R
\docType{data}
\name{taxo}
\alias{taxo}
\title{taxo}
\format{
An object of class \code{data.frame} with 2537 rows and 4 columns.
}
\source{
https://entrepot.recherche.data.gouv.fr/dataset.xhtml?persistentId=doi:10.57745/7IVO3E
}
\usage{
taxo
}
\description{
#' @format A dataframe with 2537 msps (rows) and 4 columns:
\describe{
\item{ msp_id }{  string }
\item{ species }{  string }
\item{ genus }{  string }
\item{ catalogue }{  string indicating gut and/or oral }
}
}
\keyword{datasets}
