% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truth_by_prevalence.R
\name{truth_by_prevalence}
\alias{truth_by_prevalence}
\title{Give true neighbors by level of prevalence}
\usage{
truth_by_prevalence(edge_table, prev_list)
}
\arguments{
\item{edge_table}{Dataframe. The result of prev_for_selected_nodes()}

\item{prev_list}{List of numeric. The prevalences to be studied. Required format is decimal: 0.20 for 20\% of prevalence}
}
\value{
List of dataframes. Each element of the list corresponds to a dataframe of true edges by level of prevalence
}
\description{
Give true neighbors by level of prevalence
}
\examples{
tiny_data <- data.frame(
  species = c("One bacteria", "One bacterium L", "One bacterium G", "Two bact"),
  msp_name = c("msp_1", "msp_2", "msp_3", "msp_4"),
  SAMPLE1 = c(0, 1.328425e-06, 0, 1.527688e-07),
  SAMPLE2 = c(1.251707e-07, 1.251707e-07, 3.985320e-07, 0),
  SAMPLE3 = c(0, 0, 4.926046e-09, 5.626392e-06),
  SAMPLE4 = c(0, 0, 2.98320e-05, 0)
)

tiny_graph <- graph_step(
  tiny_data,
  col_module_id = "msp_name",
  annotation_level = "species",
  seed = 20242025
) \%>\%
  suppressWarnings()

tiny_truth <- prev_for_selected_nodes(
  tiny_data,
  tiny_graph,
  col_module_id = "msp_name",
  annotation_level = "species",
  object_of_interest = "bacterium"
)

tiny_true_edges <- truth_by_prevalence(tiny_truth, c(0.20, 0.30))
}
