% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_neighborfinder.R
\name{apply_NeighborFinder}
\alias{apply_NeighborFinder}
\title{Apply NeighborFinder on raw data}
\usage{
apply_NeighborFinder(
  data_with_taxo,
  bact_of_interest,
  col_msp_id,
  taxo_level,
  data_type = "fpkm",
  prev_level = 0.3,
  filtering_top = 20,
  seed = NULL,
  ...
)
}
\arguments{
\item{data_with_taxo}{Dataframe. The abundance table merged with the species names. Required format: species are the rows and samples are the columns. The first column must be the species name, the second is the msps name, and each subsequent column is a sample}

\item{bact_of_interest}{String. The name of the bacteria or species of interest}

\item{col_msp_id}{String. The name of the column with the msp names in taxo}

\item{taxo_level}{String. The name of the column of the taxonomic level to be studied in taxo}

\item{data_type}{String. Default value is "fpkm". If your dataset is not of type "fpkm", indicate one of the following equivalent words: "comptage","couverture","coverage"}

\item{prev_level}{Numeric. The prevalence to be studied. Required format is decimal: 0.20 for 20\% of prevalence}

\item{filtering_top}{Numeric. The filtering top percentage to be studied. Required format is: 10 for top 10\%}

\item{seed}{Numeric. The seed number, ensuring reproducibility}

\item{covar}{String or formula. Formula or the name of the column of the covariate in the metadata table. Note that "study_accession" is equivalent to ~study_accession}

\item{meta_df}{Dataframe. The dataframe giving metadata information}

\item{sample_col}{String. The name of the column in metadata indicating the sample names, it should be consistent with the colnames of 'df'#'}
}
\value{
Dataframe. Returns results after using NeighborFinder(): for each msp from 'bact_of_interest', the names of their neighbors and the corresponding coefficients calculated by cv.glmnet()
}
\description{
Apply NeighborFinder on raw data
}
\examples{
data(data)
res_CRC_JPN<-apply_NeighborFinder(data$CRC_JPN, bact_of_interest="Escherichia coli", col_msp_id="msp_id", taxo_level="species", seed=20232024)
}
