% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_nf_simple.R
\name{apply_NF_simple}
\alias{apply_NF_simple}
\title{Apply NeighborFinder simplest version on raw data}
\usage{
apply_NF_simple(
  data_with_annotation,
  object_of_interest,
  col_module_id,
  annotation_level,
  data_type = "fpkm",
  prev_level = 0.3,
  filtering_top = 20,
  seed = NULL,
  ...
)
}
\arguments{
\item{data_with_annotation}{Dataframe. The abundance table merged with the module names. Required format: modules are the rows and samples are the columns. The first column must be the modules name (e.g. species), the second is the module ID (e.g. msp), and each subsequent column is a sample}

\item{object_of_interest}{String. The name of the bacteria or species of interest or a key word in the functional module definition}

\item{col_module_id}{String. The name of the column with the module names in annotation_table}

\item{annotation_level}{String. The name of the column with the level to be studied. Examples: species, genus, level_1}

\item{data_type}{String. Default value is "fpkm". If your dataset is not of type "fpkm", indicate "coverage"}

\item{prev_level}{Numeric. The prevalence to be studied. Required format is decimal: 0.20 for 20\% of prevalence}

\item{filtering_top}{Numeric. The filtering top percentage to be studied. Required format is: 10 for top 10\%}

\item{seed}{Numeric. The seed number, ensuring reproducibility}

\item{covar}{String or formula. Formula or the name of the column of the covariate in the metadata table. Note that "study_accession" is equivalent to ~study_accession}

\item{meta_df}{Dataframe. The dataframe giving metadata information}

\item{sample_col}{String. The name of the column in metadata indicating the sample names, it should be consistent with the colnames of 'df'}
}
\value{
Dataframe. Returns results after using NeighborFinder(): for each module ID from 'object_of_interest', the names of their neighbors and the corresponding coefficients calculated by cv.glmnet()
}
\description{
Apply NeighborFinder simplest version on raw data
}
\examples{
data(data)
res_CRC_JPN<-apply_NF_simple(data$CRC_JPN, object_of_interest="Escherichia coli", col_module_id="msp_id", annotation_level="species", seed=20232024)
}
