% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_params_values.R
\name{choose_params_values}
\alias{choose_params_values}
\title{Render a table to give an indication of the values to choose for the prevalence level and the top filtering percentage}
\usage{
choose_params_values(
  data_with_taxo,
  bact_of_interest,
  list_sims,
  prev_list = NULL,
  filtering_list = c(0.2),
  graph_file = NULL,
  col_msp_id,
  seed = NULL
)
}
\arguments{
\item{data_with_taxo}{Dataframe. The abundance table merged with the species names. Required format: species are the rows and samples are the columns. The first column must be the species name, the second is the msps name, and each subsequent column is a sample}

\item{bact_of_interest}{String. The name of the bacteria or species of interest}

\item{list_sims}{List of dataframes. The result of simulate_by_prevalence()}

\item{prev_list}{List of numeric. The prevalences to be studied. Required format is decimal: 0.20 for 20\% of prevalence}

\item{filtering_list}{List of numeric. The filtering top percentages to be studied. Required format is: 10 for the top 10\%}

\item{graph_file}{Dataframe. The object generated by graph_step() function}

\item{col_msp_id}{String. The name of the column with the msp names in taxo}

\item{seed}{Numeric. The seed number, ensuring reproducibility}
}
\value{
Dataframe. Returns precision and recall rates before and after using NeighborFinder()
}
\description{
Render a table to give an indication of the values to choose for the prevalence level and the top filtering percentage
}
\examples{
data(data)
data(graphs)
simulations <- simulate_by_prevalence(data$CRC_JPN, prev_list=c(0.20,0.30), graph_file=graphs$CRC_JPN, col_msp_id="msp_id", seed=20232024)
choose_params_values(data$CRC_JPN, "Klebsiella", simulations, filtering_list=c(10,20), graph_file = graphs$CRC_JPN, col_msp_id="msp_id", seed = 20232024)
}
