% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_step.R
\name{graph_step}
\alias{graph_step}
\title{Generate a graph}
\usage{
graph_step(data_with_taxo, col_msp_id, type = "fpkm")
}
\arguments{
\item{data_with_taxo}{Dataframe. The abundance table merged with the species names. Required format: species are the rows and samples are the columns. The first column must be the species name, the second is the msps name, and each subsequent column is a sample}

\item{col_msp_id}{String. The name of the column with the msp names in taxo}

\item{type}{String. Default value is "fpkm". If your dataset is not of type "fpkm", indicate one of the following equivalent words in 'type' argument: "comptage","couverture","coverage"}
}
\value{
Dataframe. The dataframe is composed of 0 and 1 corresponding to the existence of edges on the graph.
}
\description{
Generate a graph
}
\examples{
tiny_data<-data.frame(species=c("One bacteria","One bacterium L","One bacterium G","Two bact"),
                      msp_name=c("msp_1","msp_2","msp_3","msp_4"),
                      SAMPLE1=c(0,1.328425e-06,0,1.527688e-07),
                      SAMPLE2=c(1.251707e-07,1.251707e-07,3.985320e-07,0),
                      SAMPLE3=c(0,0,4.926046e-09,5.626392e-06),
                      SAMPLE4=c(0,0,2.98320e-05,0))

tiny_graph<-graph_step(tiny_data, col_msp_id="msp_name", type="fpkm") \%>\% suppressWarnings()
}
