% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_step.R
\name{graph_step}
\alias{graph_step}
\title{Generate a graph}
\usage{
graph_step(
  data_with_annotation,
  col_module_id,
  annotation_level,
  type = "fpkm"
)
}
\arguments{
\item{data_with_annotation}{Dataframe. The abundance table merged with the module names. Required format: modules are the rows and samples are the columns. The first column must be the modules name (e.g. species), the second is the module ID (e.g. msp), and each subsequent column is a sample}

\item{col_module_id}{String. The name of the column with the module names in the annotation table}

\item{annotation_level}{String. The name of the column with the level to be studied. Examples: species, genus, level_1}

\item{type}{String. Default value is "fpkm". If your dataset is not of type "fpkm", indicate in 'type' argument: "coverage"}
}
\value{
Dataframe. The dataframe is composed of 0 and 1 corresponding to the existence of edges on the graph.
}
\description{
Generate a graph
}
\examples{
tiny_data<-data.frame(species=c("One bacteria","One bacterium L","One bacterium G","Two bact"),
                      msp_name=c("msp_1","msp_2","msp_3","msp_4"),
                      SAMPLE1=c(0,1.328425e-06,0,1.527688e-07),
                      SAMPLE2=c(1.251707e-07,1.251707e-07,3.985320e-07,0),
                      SAMPLE3=c(0,0,4.926046e-09,5.626392e-06),
                      SAMPLE4=c(0,0,2.98320e-05,0))

tiny_graph<-graph_step(tiny_data, col_module_id="msp_name", annotation_level="species", type="fpkm") \%>\% suppressWarnings()
}
