% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs-for-crc-patients.R
\docType{data}
\name{graphs}
\alias{graphs}
\title{graphs}
\format{
A list of dataframes corresponding to graphs based on synthetic data.
\describe{
\item{ CRC_JPN }{  graph corresponding to data with only Japanese patients diagnosed with colorectal cancer }
\item{ CRC_JPN_CHN_EUR }{  graph corresponding to data gathering Japanese, Chinese and European patients diagnosed with colorectal cancer }
}
}
\usage{
graphs
}
\description{
graphs
}
\keyword{datasets}
