% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_module.R
\name{identify_module}
\alias{identify_module}
\title{List the modules corresponding to a given object of interest}
\usage{
identify_module(
  object_of_interest,
  annotation_table,
  col_module_id,
  annotation_level = "species"
)
}
\arguments{
\item{object_of_interest}{String. The name of the bacteria or species of interest or a key word in the functional module definition}

\item{annotation_table}{Dataframe. The dataframe gathering the taxonomic or functional module correspondence information}

\item{col_module_id}{String. The name of the column with the module names in the annotation table}

\item{annotation_level}{String. The name of the column with the level to be studied. Examples: species, genus, level_1. Default value is set to the species level}
}
\value{
List of string. The module names of the corresponding object of interest
}
\description{
List the modules corresponding to a given object of interest
}
\examples{
df_taxo <- data.frame(
  msp_name = c("msp_1", "msp_2", "msp_3", "msp_4"),
  genus = c("One", "One", "One", "Two"),
  species = c("One bacteria", "One bacterium L", "One bacterium G", "Two bact")
)
identify_module(object_of_interest = "bacterium", annotation_table = df_taxo, col_module_id = "msp_name", annotation_level = "species")
identify_module(object_of_interest = "One", annotation_table = df_taxo, col_module_id = "msp_name", annotation_level = "species")
}
