% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersections_network.R
\name{intersections_network}
\alias{intersections_network}
\title{Display the intersection network from 2 or more datasets}
\usage{
intersections_network(
  res_list,
  threshold,
  taxo,
  col_msp_id,
  taxo_level,
  bact_of_interest,
  taxo_option = FALSE,
  node_size = 12,
  label_size = 4,
  species_color = "cadetblue2",
  seed = NULL
)
}
\arguments{
\item{res_list}{List of dataframes. The results from apply_NeighborFinder() on several datasets}

\item{threshold}{Numeric. Integer corresponding to the minimum number of datasets in which you want neighbors to have been found}

\item{taxo}{Dataframe. The dataframe gathering the taxonomic correspondence information}

\item{col_msp_id}{String. The name of the column with the msp names in taxo}

\item{taxo_level}{String. The name of the column of the taxonomic level to be studied in taxo}

\item{bact_of_interest}{String. The name of the bacteria or species of interest}

\item{taxo_option}{Boolean. Default value is False. If True: labels on nodes become species names instead of msps names}

\item{node_size}{Numeric. The parameter to adjust size of nodes}

\item{label_size}{Numeric. The parameter to adjust size of labels}

\item{species_color}{String. The name of the color to differentiate the nodes corresponding to 'bact_of_interest' from the other msps}

\item{seed}{Numeric. The seed number, ensuring reproducibility}
}
\value{
Network. Visualization of NeighborFinder results from several datasets
}
\description{
Display the intersection network from 2 or more datasets
}
\examples{
data(taxo)
data(data)
data(metadata)
res_CRC_JPN<-apply_NeighborFinder(data$CRC_JPN, bact_of_interest="Escherichia coli", col_msp_id="msp_id", taxo_level="species", seed=20232024)
res_CRC_CHN<-apply_NeighborFinder(data$CRC_CHN, bact_of_interest="Escherichia coli", col_msp_id="msp_id", taxo_level="species", seed=20232024, covar= ~ study_accession, meta_df=metadata$CRC_CHN, sample_col="secondary_sample_accession")
res_CRC_EUR<-apply_NeighborFinder(data$CRC_EUR, bact_of_interest="Escherichia coli", col_msp_id="msp_id", taxo_level="species", seed=20232024, covar= ~ study_accession, meta_df=metadata$CRC_EUR, sample_col="secondary_sample_accession")

intersections_network(res_list=list(res_CRC_JPN, res_CRC_CHN, res_CRC_EUR), taxo, threshold=2, "Escherichia coli", col_msp_id="msp_id", taxo_level="species", label_size=5, taxo_option=TRUE, seed=3)
}
