% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersections_table.R
\name{intersections_table}
\alias{intersections_table}
\title{Display the intersection table summarizing the results from 2 or more datasets}
\usage{
intersections_table(res_list, threshold, taxo, col_msp_id, bact_of_interest)
}
\arguments{
\item{res_list}{List of dataframes. The results from apply_NeighborFinder() on several datasets}

\item{threshold}{Numeric. Integer corresponding to the minimum number of datasets in which you want neighbors to have been found}

\item{taxo}{Dataframe. The dataframe gathering the taxonomic correspondence information}

\item{col_msp_id}{String. The name of the column with the msp names in taxo}

\item{bact_of_interest}{String. The name of the bacteria or species of interest}
}
\value{
Dataframe. Table gathering the intersection of NeighborFinder results from several datasets. The column 'datasets' indicates the datasets in which the same neighbor has been found, the column 'intersections' indicates the number of datasets in which the same neighbor has been found
}
\description{
Display the intersection table summarizing the results from 2 or more datasets
}
\examples{
data(taxo)
data(data)
data(metadata)
res_CRC_JPN<-apply_NeighborFinder(data$CRC_JPN, bact_of_interest="Escherichia coli", col_msp_id="msp_id", taxo_level="species", seed=20232024)
res_CRC_CHN<-apply_NeighborFinder(data$CRC_CHN, bact_of_interest="Escherichia coli", col_msp_id="msp_id", taxo_level="species", seed=20232024, covar= ~ study_accession, meta_df=metadata$CRC_CHN, sample_col="secondary_sample_accession")
res_CRC_EUR<-apply_NeighborFinder(data$CRC_EUR, bact_of_interest="Escherichia coli", col_msp_id="msp_id", taxo_level="species", seed=20232024, covar= ~ study_accession, meta_df=metadata$CRC_EUR, sample_col="secondary_sample_accession")

intersections_table(res_list=list(res_CRC_JPN, res_CRC_CHN, res_CRC_EUR), threshold=2, taxo, col_msp_id="msp_id", "Escherichia coli")
}
