% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_for_selected_nodes.R
\name{prev_for_selected_nodes}
\alias{prev_for_selected_nodes}
\title{Extract edges in graph involving any msp in bact_of_interest set}
\usage{
prev_for_selected_nodes(
  data_with_taxo,
  graph_file,
  col_msp_id,
  bact_of_interest = NULL
)
}
\arguments{
\item{data_with_taxo}{Dataframe. The abundance table merged with the species names. Required format: species are the rows and samples are the columns. The first column must be the species name, the second is the msps name, and each subsequent column is a sample}

\item{graph_file}{Dataframe. The object generated by graph_step() function}

\item{col_msp_id}{String. The name of the column with the msp names in taxo}

\item{bact_of_interest}{String. The name of the bacteria or species of interest}
}
\value{
Dataframe. The dataframe of edges in the graph involving msps corresponding to bact_of_interest and their corresponding prevalences.
}
\description{
Extract edges in graph involving any msp in bact_of_interest set
}
\examples{
tiny_data<-data.frame(species=c("One bacteria","One bacterium L","One bacterium G","Two bact"),
                      msp_name=c("msp_1","msp_2","msp_3","msp_4"),
                      SAMPLE1=c(0,1.328425e-06,0,1.527688e-07),
                      SAMPLE2=c(1.251707e-07,1.251707e-07,3.985320e-07,0),
                      SAMPLE3=c(0,0,4.926046e-09,5.626392e-06),
                      SAMPLE4=c(0,0,2.98320e-05,0))
tiny_graph<-graph_step(tiny_data, col_msp_id="msp_name", type="fpkm") \%>\% suppressWarnings()

tiny_truth <-prev_for_selected_nodes(tiny_data, tiny_graph, col_msp_id="msp_name", bact_of_interest="bacterium")
}
