% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_by_prevalence.R
\name{simulate_by_prevalence}
\alias{simulate_by_prevalence}
\title{List the simulated count tables by level of prevalence}
\usage{
simulate_by_prevalence(
  data_with_taxo,
  prev_list,
  graph_file = NULL,
  col_msp_id,
  sample_size = 500,
  seed
)
}
\arguments{
\item{data_with_taxo}{Dataframe. The abundance table merged with the species names. Required format: species are the rows and samples are the columns. The first column must be the species name, the second is the msps name, and each subsequent column is a sample}

\item{prev_list}{List of numeric. The prevalences to be studied. Required format is decimal: 0.20 for 20\% of prevalence.}

\item{graph_file}{Dataframe. The object generated by graph_step() function}

\item{col_msp_id}{String. The name of the column with the msp names in data_with_taxo}

\item{sample_size}{Numeric. The size to be considerated, the value of 500 is recommended}

\item{seed}{Numeric. The seed number, ensuring reproducibility}
}
\value{
List of dataframes. Each element of the list corresponds to a level of prevalence and is a simulated abundance table
}
\description{
List the simulated count tables by level of prevalence
}
\examples{
tiny_data<-data.frame(species=c("One bacteria","One bacterium L","One bacterium G","Two bact"),
                      msp_name=c("msp_1","msp_2","msp_3","msp_4"),
                      SAMPLE1=c(0,1.328425e-06,0,1.527688e-07),
                      SAMPLE2=c(1.251707e-07,1.251707e-07,3.985320e-07,0),
                      SAMPLE3=c(0,0,4.926046e-09,5.626392e-06),
                      SAMPLE4=c(0,0,2.98320e-05,0))

tiny_graph<-graph_step(tiny_data, col_msp_id="msp_name", type="fpkm") \%>\% suppressWarnings()

tiny_sims<-simulate_by_prevalence(tiny_data, prev_list=c(0.20,0.30), graph_file=tiny_graph, col_msp_id="msp_name", sample_size=500, seed=20232024)
}
