% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_from_ecdf.R
\name{simulate_from_ecdf}
\alias{simulate_from_ecdf}
\title{Simulate data (extracted from OneNet package)
Generates synthetic count data based on empirical cumulative distribution (ecdf) of real count data}
\usage{
simulate_from_ecdf(real_data, Sigma, n, seed = 10010, verbatim = FALSE)
}
\arguments{
\item{real_data}{Matrix. Contains real count data of size n by p}

\item{Sigma}{Matrix. Covariance structure of size p by p}

\item{n}{Numeric. Number of samples}

\item{seed}{Numeric. Seed number for data generation}

\item{verbatim}{Boolean. If TRUE: iteration and index calculation for each step are printed out}
}
\value{
Matrix. The vector from the upper triangular part of A.mat
}
\description{
Simulate data (extracted from OneNet package)
Generates synthetic count data based on empirical cumulative distribution (ecdf) of real count data
}
