% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_filter.R
\name{test_filter}
\alias{test_filter}
\title{Render a table gathering precision and recall rates before and after filtering on coefficient values}
\usage{
test_filter(df_before, df_after, prevs = NULL)
}
\arguments{
\item{df_before}{Dataframe. The one of true neighbors}

\item{df_after}{Dataframe. The one of detected neighbors}

\item{prevs}{List of numeric. The prevalences to be studied. Required format is decimal: 0.20 for 20\% of prevalence}
}
\value{
Dataframe. Returns the precision and recall rates before and after the modification
}
\description{
Render a table gathering precision and recall rates before and after filtering on coefficient values
}
\examples{
data(data)
data(taxo)
data(graphs)
#Generate dataframe with true neighbors
df_true<-truth_by_prevalence(edge_table = prev_for_selected_nodes(data$CRC_JPN, graphs$CRC_JPN, "msp_id", "Klebsiella"), prev_list=c(0.20,0.30))
#Generate dataframe with detected neighbors
normed_JPN<-norm_data(data$CRC_JPN, col_msp_id="msp_id", prev_list=c(0.20, 0.25, 0.30), type="fpkm")
df_detected<-cvglm_to_coeffs_by_bact(list_dfs=normed_JPN, test_msp=identify_msp("Klebsiella",taxo,"msp_id"), seed=20232024)
#Use final_step() to gather both
neighbors<-final_step(df_true, df_detected)

neighbors_modif<-final_step(df_true, res_by_filtering(df_detected, filtering_list=c(10,50)))

scores<-test_filter(neighbors, neighbors_modif, prevs=names(df_true))
}
