% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_network.R
\name{visualize_network}
\alias{visualize_network}
\title{Display network after applying NeighborFinder}
\usage{
visualize_network(
  res_NeighborFinder,
  taxo,
  col_msp_id,
  taxo_level,
  bact_of_interest,
  taxo_option = FALSE,
  node_size = 12,
  label_size = 4,
  species_color = "cadetblue2",
  seed = NULL
)
}
\arguments{
\item{res_NeighborFinder}{Dataframe. The result from apply_NeighborFinder()}

\item{taxo}{Dataframe. The dataframe gathering the taxonomic correspondence information}

\item{col_msp_id}{String. The name of the column with the msp names in taxo}

\item{taxo_level}{String. The name of the column of the taxonomic level to be studied in taxo}

\item{bact_of_interest}{String. The name of the bacteria or species of interest}

\item{taxo_option}{Boolean. Default value is False. If True: labels on nodes become species names instead of msps names}

\item{node_size}{Numeric. The parameter to adjust size of nodes}

\item{label_size}{Numeric. The parameter to adjust size of labels}

\item{species_color}{String. The name of the color to differentiate the nodes corresponding to 'bact_of_interest' from the other msps}

\item{seed}{Numeric. The seed number, ensuring reproducibility}
}
\value{
Network. Visualization of NeighborFinder results
}
\description{
Display network after applying NeighborFinder
}
\examples{
data(taxo)
data(data)
res_CRC_JPN<-apply_NeighborFinder(data$CRC_JPN, bact_of_interest="Escherichia coli", col_msp_id="msp_id", taxo_level="species", seed=20232024)

visualize_network(res_CRC_JPN, taxo, bact_of_interest="Escherichia coli", col_msp_id="msp_id", taxo_level="species", label_size=5)
#With species names instead of msps names
visualize_network(res_CRC_JPN, taxo, bact_of_interest="Escherichia coli", col_msp_id="msp_id", taxo_level="species", label_size=5, taxo_option=TRUE, seed=2)
#With esthetic changes
visualize_network(res_CRC_JPN, taxo, bact_of_interest="Escherichia coli", col_msp_id="msp_id", taxo_level="species", taxo_option=TRUE, node_size=15, label_size=6, species_color= "orange", seed=2)
}
